requirements.txt
mcp>=1.0.0
anthropic>=0.40.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0
sqlparse>=0.5.0
pgvector>=0.3.0
voyageai>=0.3.0